requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
python-dotenv
pydantic
tiktoken
chromadb
mlflow